refactor(tools-config): unify config types and restructure modules
- Consolidate and simplify shared tool config types
- Remove redundant internal types and outdated unit tests
- Rename validator directories and clean up structure
- Improve type safety and reuse across validation logic

- 内容が重複していたユニットテストのうち、古いユニットテストを削除
