--- packages/@esta-core/tools-config/shared/schemas/index.ts
@@ -1,9 +1,0 @@
-// @(#) : スキーマ定義のエクスポート
-//
-// Copyright (c) 2025 atsushifx <http://github.com/atsushifx>
-//
-// This software is released under the MIT License.
-// https://opensource.org/licenses/MIT
-
-// Tools schemas (consolidated from src/schemas)
-export { CompleteToolsConfigSchema, ToolEntrySchema, ToolsConfigSchema } from './tools.schemas';
