+++ packages/@esta-core/tools-config/shared/types/index.ts
@@ -1,1 +1,14 @@
-export * from './tools.types';
+// shared/types/index.ts
+// @(#) : 共有型定義のエクスポート
+//
+// Copyright (c) 2025 atsushifx <http://github.com/atsushifx>
+//
+// This software is released under the MIT License.
+// https://opensource.org/licenses/MIT
+
+// Tools configuration types
+export type {
+  PartialToolsConfig,
+  ToolEntry,
+  ToolsConfig,
+} from './toolsConfig.types';
