--- packages/@esta-core/tools-config/shared/types/toolEntry.types.ts
@@ -1,9 +1,0 @@
-/**
- * ツール設定エントリー
- */
-export type ToolEntry = {
-  installer: string;
-  id: string;
-  repository: string;
-  options?: Record<string, unknown>;
-};
