+++ packages/@esta-core/tools-config/src/core/config/index.ts
@@ -7,9 +7,3 @@
 // https://opensource.org/licenses/MIT
 
 export { loadConfig } from './loader';
-
-export {
-  getDefaultTool,
-  getDefaultTools,
-  getDefaultToolsConfig,
-} from './defaults';
