+++ packages/@esta-core/tools-config/src/defaults.ts
@@ -1,4 +1,4 @@
-// src/core/config/defaults.ts
+// src/defaults.ts
 // @(#) : デフォルト設定提供機能
 //
 // Copyright (c) 2025 atsushifx <http://github.com/atsushifx>
@@ -7,9 +7,9 @@
 // https://opensource.org/licenses/MIT
 
 import { parse } from 'valibot';
-import { DEFAULT_INSTALL_DIR, DEFAULT_TEMP_DIR } from '../../internal/constants';
-import { CompleteToolsConfigSchema } from '../../internal/schemas';
-import type { ToolEntry, ToolsConfig } from '../../internal/types';
+import { DEFAULT_INSTALL_DIR, DEFAULT_TEMP_DIR } from './internal/constants';
+import { CompleteToolsConfigSchema } from './internal/schemas';
+import type { ToolEntry, ToolsConfig } from './internal/types';
 
 /**
  * デフォルトのツール設定リスト
@@ -67,7 +67,7 @@ const defaultTools: ToolEntry[] = [
  * デフォルトのツール設定を取得する（validated済み）
  * @returns 検証済みのToolsConfig
  */
-export const getDefaultToolsConfig = (): ToolsConfig => {
+export const defaultToolsConfig = (): ToolsConfig => {
   const config = {
     defaultInstallDir: DEFAULT_INSTALL_DIR,
     defaultTempDir: DEFAULT_TEMP_DIR,
@@ -76,32 +76,4 @@ export const getDefaultToolsConfig = (): ToolsConfig => {
 
   // スキーマで検証・正規化して返す
   return parse(CompleteToolsConfigSchema, config) as ToolsConfig;
-};
-
-/**
- * デフォルトツールリストを取得する（validated済み）
- * @returns 検証済みのツールエントリー配列
- */
-export const getDefaultTools = (): ToolEntry[] => {
-  // 各ツールエントリーを検証して返す
-  return defaultTools.map((tool) => {
-    // ToolEntryの検証は既にdefaultToolsで行われているが、
-    // 将来的にスキーマ検証を追加する場合はここで実装
-    return { ...tool };
-  });
-};
-
-/**
- * 特定のデフォルトツールを取得する（validated済み）
- * @param id ツールID
- * @returns 見つかった検証済みツールエントリー、または undefined
- */
-export const getDefaultTool = (id: string): ToolEntry | undefined => {
-  const tool = defaultTools.find((tool) => tool.id === id);
-  if (!tool) {
-    return undefined;
-  }
-
-  // ツールエントリーのコピーを返す
-  return { ...tool };
 };
