+++ packages/@esta-core/tools-config/src/internal/types/config.ts
@@ -7,30 +7,6 @@
 // https://opensource.org/licenses/MIT
 
 /**
- * 設定ファイル情報
- */
-export type ConfigFileInfo = {
-  /** ファイルパス */
-  path: string;
-  /** ファイル形式 */
-  format: SupportedConfigFileFormat;
-  /** ファイルの存在確認済みかどうか */
-  exists: boolean;
-};
-
-/**
- * 設定検索オプション
- */
-export type ConfigSearchOptions = {
-  /** 検索を開始するディレクトリ */
-  startDir?: string;
-  /** 検索するファイル名（拡張子なし） */
-  configName?: string;
-  /** 検索を停止するディレクトリ */
-  stopDir?: string;
-};
-
-/**
  * 設定読み込み結果
  */
 export type LoadConfigResult<T = unknown> = {
@@ -38,20 +14,6 @@ export type LoadConfigResult<T = unknown> = {
   success: boolean;
   /** 設定データ */
   config?: T;
-  /** 設定ファイル情報 */
-  fileInfo?: ConfigFileInfo;
   /** エラーメッセージ */
   error?: string;
 };
-
-/**
- * サポートされる設定ファイル形式
- */
-export type SupportedConfigFileFormat =
-  | 'json' // 標準JSON形式
-  | 'jsonc' // コメント付きJSON形式
-  | 'yaml' // YAML形式
-  | 'yml' // YAML形式（短縮版）
-  | 'js' // JavaScript設定ファイル
-  | 'ts' // TypeScript設定ファイル
-  | ''; // 拡張子なし（JSONとして処理）
