+++ packages/@esta-core/tools-config/src/tools-validator/index.ts
@@ -1,10 +1,10 @@
-// src/validator/index.ts
-// @(#) : 検証機能のエクスポート
+// src/toolsValidator/index.ts
+// @(#) : ツール検証機能のエクスポート
 //
 // Copyright (c) 2025 atsushifx <http://github.com/atsushifx>
 //
 // This software is released under the MIT License.
 // https://opensource.org/licenses/MIT
 
-// Tools validation
-export * from '../toolsValidator';
+// Validators
+export * from './validator';
